Move drop-card behaviour into verify endpoint
